+++ java/spark-1.5.2/utils.scala
@@ -1,6 +1,7 @@
 package sparklyr
 
 import java.io._
+import java.net.{InetAddress, ServerSocket}
 import java.util.Arrays
 
 import scala.util.Try
@@ -427,6 +428,33 @@ object Utils {
       if (cl == null) Nil else cl :: supers(cl.getSuperclass)
     }
   supers(obj.getClass).map(if (simpleName) _.getSimpleName else _.getName).toArray
+  }
+
+  def portIsAvailable(port: Int, inetAddress: InetAddress) = {
+    var ss: ServerSocket = null
+    var available = false
+
+    Try {
+        ss = new ServerSocket(port, 1, inetAddress)
+        available = true
+    }
+
+    if (ss != null) {
+        Try {
+            ss.close();
+        }
+    }
+
+    available
+  }
+
+  def nextPort(port: Int, inetAddress: InetAddress) = {
+    var freePort = port
+    while (!portIsAvailable(freePort, inetAddress) && freePort - port > 100)
+      freePort += 1
+
+    // give up after 100 port searches
+    if (freePort - port > 100) 0 else freePort;
   }
 }
